refactoring rockettile and adding helper methods to the OM* case classes
