+++ src/main/scala/diplomaticobjectmodel/model/OMISA.scala
@@ -2,6 +2,9 @@
 
 package freechips.rocketchip.diplomaticobjectmodel.model
 
+import freechips.rocketchip.rocket.RocketCoreParams
+import freechips.rocketchip.tile.XLen
+
 trait OMExtensionType extends OMEnum
 case object M extends OMExtensionType
 case object A extends OMExtensionType
@@ -36,3 +39,44 @@ case class OMISA(
   addressTranslationModes: Seq[OMAddressTranslationMode],
   _types: Seq[String] = Seq("OMISA", "OMCompoundType")
 ) extends OMCompoundType
+
+object OMISA {
+  // TODO the package object util has this code but is not seen in this scope
+  implicit class BooleanToAugmentedBoolean(val x: Boolean) extends AnyVal {
+    // this one's snagged from scalaz
+    def option[T](z: => T): Option[T] = if (x) Some(z) else None
+  }
+
+  def isa(coreParams: RocketCoreParams, xLen: Int): OMISA = {
+    val baseInstructionSet = xLen match {
+      case 32 => if (XLen == 32) RV32E else RV32I
+      case 64 => RV64I
+      case _ => throw new IllegalArgumentException(s"ERROR: Invalid Xlen: $xLen")
+    }
+
+    val isaExtSpec = ISAExtensions.specVersion _
+
+    val baseSpec = BaseExtensions.specVersion _
+
+    val baseISAVersion = baseInstructionSet match {
+      case RV32E => "1.9"
+      case RV32I => "2.0"
+      case RV64I => "2.0"
+      case _ => throw new IllegalArgumentException(s"ERROR: Invalid baseISAVersion: $baseInstructionSet")
+    }
+
+    OMISA(
+      xLen = xLen,
+      baseSpecification = baseSpec(baseInstructionSet, baseISAVersion),
+      base = baseInstructionSet,
+      m = coreParams.mulDiv.map(x => isaExtSpec(M, "2.0")),
+      a = coreParams.useAtomics.option(isaExtSpec(A, "2.0")),
+      f = coreParams.fpu.map(x => isaExtSpec(F, "2.0")),
+      d = coreParams.fpu.filter(_.fLen > 32).map(x => isaExtSpec(D, "2.0")),
+      c = coreParams.useCompressed.option(isaExtSpec(C, " 2.0")),
+      u = coreParams.useUser.option(isaExtSpec(U, "1.10")),
+      s = coreParams.useVM.option(isaExtSpec(S, "1.10")),
+      addressTranslationModes = Nil
+    )
+  }
+}
